bot/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
psycopg2-binary==2.9.1
pyTelegramBotAPI==3.8.1
requests==2.26.0
urllib3==1.26.6
